Location data can be toggled off or on with keyboard input
Currently the location data can be turned on or off by pressing E key on the keyboard. The process is a little buggy as multiple frames can render during the time of the button press, causing the menu to glitch off or on too fast. As of now, no actual data is connected to panel, need global variables to access for the information.